feat: revert moving grammar to separate folder
