+++ src/abi/AbiFunction.ts
@@ -2,7 +2,7 @@ import { AstExpression } from "../grammar/ast";
 import { CompilerContext } from "../context";
 import { WriterContext } from "../generator/Writer";
 import { TypeRef } from "../types/types";
-import { SrcInfo } from "../grammar/prev";
+import { SrcInfo } from "../grammar";
 
 export type AbiFunction = {
     name: string;
