+++ src/errors.ts
@@ -2,7 +2,7 @@ import { MatchResult } from "ohm-js";
 import path from "path";
 import { cwd } from "process";
 import { AstFuncId, AstId, AstTypeId } from "./grammar/ast";
-import { ItemOrigin, SrcInfo } from "./grammar/prev";
+import { ItemOrigin, SrcInfo } from "./grammar";
 
 export class TactError extends Error {
     readonly loc?: SrcInfo;
