+++ src/generator/writers/writeSerialization.ts
@@ -1,6 +1,6 @@
 import { contractErrors } from "../../abi/errors";
 import { throwInternalCompilerError } from "../../errors";
-import { dummySrcInfo, ItemOrigin } from "../../grammar/prev";
+import { dummySrcInfo, ItemOrigin } from "../../grammar";
 import { AllocationCell, AllocationOperation } from "../../storage/operation";
 import { StorageAllocation } from "../../storage/StorageAllocation";
 import { getType } from "../../types/resolveDescriptors";
