+++ src/optimizer/test/partial-eval.spec.ts
@@ -12,7 +12,7 @@ import { CompilerContext } from "../../context";
 import { ExpressionTransformer, Rule } from "../types";
 import { AssociativeRule3 } from "../associative";
 import { evalBinaryOp, evalUnaryOp } from "../../interpreter";
-import { getParser } from "../../grammar/prev";
+import { getParser } from "../../grammar";
 
 const MAX: string =
     "115792089237316195423570985008687907853269984665640564039457584007913129639935";
