+++ src/types/resolveStatements.ts
@@ -30,7 +30,7 @@ import { FunctionDescription, printTypeRef, TypeRef } from "./types";
 import { evalConstantExpression } from "../constEval";
 import { ensureInt } from "../interpreter";
 import { crc16 } from "../utils/crc16";
-import { SrcInfo } from "../grammar/prev";
+import { SrcInfo } from "../grammar";
 
 export type StatementContext = {
     root: SrcInfo;
